move test back in original location
