+++ test/e2e/nstemplatetier_test.go
@@ -3,10 +3,12 @@ package e2e
 import (
 	"context"
 	"fmt"
+	"strings"
 	"testing"
 
 	toolchainv1alpha1 "github.com/codeready-toolchain/api/pkg/apis/toolchain/v1alpha1"
 	"github.com/codeready-toolchain/toolchain-e2e/testsupport"
+	"github.com/codeready-toolchain/toolchain-e2e/tiers"
 	. "github.com/codeready-toolchain/toolchain-e2e/wait"
 	"github.com/operator-framework/operator-sdk/pkg/test"
 	"github.com/stretchr/testify/assert"
@@ -28,7 +30,7 @@ const (
 	MaxPoolSize = 5 // same as hard-coded value in host operator
 )
 
-func TestPromoteToOtherTier(t *testing.T) {
+func TestNSTemplateTiers(t *testing.T) {
 	// given
 	tierList := &toolchainv1alpha1.NSTemplateTierList{}
 	ctx, awaitility := testsupport.WaitForDeployments(t, tierList)
@@ -88,6 +90,189 @@ func TestPromoteToOtherTier(t *testing.T) {
 		err := hostAwaitility.WaitUntilChangeTierRequestDeleted(name)
 		assert.NoError(t, err)
 	}
+}
+
+func TestUpdateToOtherRevisionOfTierTemplates(t *testing.T) {
+	// given
+	tierList := &toolchainv1alpha1.NSTemplateTierList{}
+	ctx, awaitility := testsupport.WaitForDeployments(t, tierList)
+	defer ctx.Cleanup()
+
+	// Create and approve "testing-template-updates" user
+	murName := "testingtemplateupdates"
+	userSignup := createAndApproveSignup(t, awaitility, murName)
+
+	tierTemplatesByTier := getTierTemplatesMappedByTier(t, awaitility)
+	templatesFromAdvanced := createBasicTierTemplate(t, ctx, awaitility, tierTemplatesByTier["advanced"], "123abcd")
+	templatesFromTeam := createBasicTierTemplate(t, ctx, awaitility, tierTemplatesByTier["team"], "456efgh")
+	templatesFromBasic := createBasicTierTemplate(t, ctx, awaitility, tierTemplatesByTier["basic"], "789ijkl")
+
+	// wait for the user to be provisioned for the first time
+	verifyResourcesProvisionedForSignup(t, awaitility, userSignup, "basic")
+
+	t.Run("update to refs basic-*-123abcd made out of the advanced TierTemplates", func(t *testing.T) {
+		// when update to basic-*-123abcd
+		replaceMurRefs(t, awaitility, murName, templatesFromAdvanced)
+
+		// then
+		verifyMurIsProvisionedForRevisionMadeOutOfOtherTier(t, awaitility, murName, "advanced", "123abcd")
+	})
+
+	t.Run("update to refs basic-*-456efgh made out of the team TierTemplates", func(t *testing.T) {
+		// when update to basic-*-456efgh
+		replaceMurRefs(t, awaitility, murName, templatesFromTeam)
+
+		// then
+		verifyMurIsProvisionedForRevisionMadeOutOfOtherTier(t, awaitility, murName, "team", "456efgh")
+	})
+
+	t.Run("update to refs basic-*-789ijkl made out of the basic TierTemplates", func(t *testing.T) {
+		// when update to basic-*-789ijkl
+		replaceMurRefs(t, awaitility, murName, templatesFromBasic)
+
+		// then
+		verifyMurIsProvisionedForRevisionMadeOutOfOtherTier(t, awaitility, murName, "basic", "789ijkl")
+	})
+}
+
+func TestTierTemplates(t *testing.T) {
+	// given
+	tierList := &toolchainv1alpha1.NSTemplateTierList{}
+	ctx, awaitility := testsupport.WaitForDeployments(t, tierList)
+	defer ctx.Cleanup()
+	// when the tiers are created during the startup then we can verify them
+	allTiers := &toolchainv1alpha1.TierTemplateList{}
+	err := awaitility.Client.List(context.TODO(), allTiers, client.InNamespace(awaitility.HostNs))
+	// verify that we have 11 tier templates (4+4+3)
+	require.NoError(t, err)
+	assert.Len(t, allTiers.Items, 11)
+}
+
+func TestUpdateOfNamespacesWithLegacyLabels(t *testing.T) {
+	// given
+	tierList := &toolchainv1alpha1.NSTemplateTierList{}
+	ctx, awaitility := testsupport.WaitForDeployments(t, tierList)
+	defer ctx.Cleanup()
+	for _, nsType := range []string{"code", "dev", "stage"} {
+		err := awaitility.Client.Create(context.TODO(), &corev1.Namespace{
+			ObjectMeta: metav1.ObjectMeta{
+				Name: "legacy-" + nsType,
+				Labels: map[string]string{
+					"toolchain.dev.openshift.com/provider": "codeready-toolchain",
+					"toolchain.dev.openshift.com/owner":    "legacy",
+					"toolchain.dev.openshift.com/tier":     "basic",
+					"toolchain.dev.openshift.com/type":     nsType,
+				},
+			},
+		}, &test.CleanupOptions{})
+		require.NoError(t, err)
+	}
+
+	// when
+	legacySignup := createAndApproveSignup(t, awaitility, "legacy")
+
+	// then
+	verifyResourcesProvisionedForSignup(t, awaitility, legacySignup, "basic")
+}
+
+func newChangeTierRequest(namespace, tier, murName string) *toolchainv1alpha1.ChangeTierRequest {
+	return &toolchainv1alpha1.ChangeTierRequest{
+		ObjectMeta: metav1.ObjectMeta{
+			Namespace:    namespace,
+			GenerateName: "changetierrequest-",
+		},
+		Spec: toolchainv1alpha1.ChangeTierRequestSpec{
+			TierName: tier,
+			MurName:  murName,
+		},
+	}
+}
+
+func verifyMurIsProvisionedForRevisionMadeOutOfOtherTier(t *testing.T, awaitility *Awaitility, murName, originalTier, revision string) {
+	_, err := awaitility.Host().WaitForMasterUserRecord(murName, UntilMasterUserRecordHasConditions(provisioned(), provisionedNotificationCRCreated()))
+	require.NoError(t, err)
+
+	memberAwait := awaitility.Member()
+	userAccount, err := memberAwait.WaitForUserAccount(murName)
+	require.NoError(t, err)
+	templateRefs := tiers.GetTemplateRefsForTierAndRevision(awaitility.Host(), "basic", revision)
+
+	nsTemplateSet, err := memberAwait.WaitForNSTmplSet(userAccount.Name, UntilNSTemplateSetHasTier("basic"))
+	assert.NoError(t, err)
+
+	tierChecks, err := tiers.NewChecks(originalTier)
+	require.NoError(t, err)
+
+	tiers.VerifyGivenNsTemplateSet(t, memberAwait, nsTemplateSet, tierChecks, templateRefs)
+}
+
+func getTierTemplatesMappedByTier(t *testing.T, awaitility *Awaitility) map[string][]*toolchainv1alpha1.TierTemplate {
+	allTiers := &toolchainv1alpha1.NSTemplateTierList{}
+	err := awaitility.Client.List(context.TODO(), allTiers, client.InNamespace(awaitility.HostNs))
+	require.NoError(t, err)
+	mappedTierTemplates := map[string][]*toolchainv1alpha1.TierTemplate{}
+	for _, tier := range allTiers.Items {
+		for _, ns := range tier.Spec.Namespaces {
+			tierTemplate, err := awaitility.Host().WaitForTierTemplate(ns.TemplateRef)
+			require.NoError(t, err)
+			mappedTierTemplates[tier.Name] = append(mappedTierTemplates[tier.Name], tierTemplate)
+		}
+		if tier.Spec.ClusterResources != nil {
+			tierTemplate, err := awaitility.Host().WaitForTierTemplate(tier.Spec.ClusterResources.TemplateRef)
+			require.NoError(t, err)
+			mappedTierTemplates[tier.Name] = append(mappedTierTemplates[tier.Name], tierTemplate)
+		}
+	}
+	return mappedTierTemplates
+}
+
+func createBasicTierTemplate(t *testing.T, ctx *test.Context, awaitility *Awaitility, tierTemplates []*toolchainv1alpha1.TierTemplate, revision string) []string {
+	var refs []string
+	for _, tierTemplate := range tierTemplates {
+		_, typeName, _, err := Split(tierTemplate.Name)
+		require.NoError(t, err)
+		ref := strings.ToLower(fmt.Sprintf("%s-%s-%s", "basic", typeName, revision))
+		newTierTemplate := &toolchainv1alpha1.TierTemplate{
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      ref,
+				Namespace: awaitility.HostNs,
+			},
+			Spec: toolchainv1alpha1.TierTemplateSpec{
+				Type:     typeName,
+				TierName: "basic",
+				Revision: revision,
+				Template: tierTemplate.Spec.Template,
+			},
+		}
+
+		newTierTemplate.Name = ref
+		refs = append(refs, ref)
+		err = awaitility.Client.Create(context.TODO(), newTierTemplate, testsupport.CleanupOptions(ctx))
+		require.NoError(t, err)
+	}
+	return refs
+}
+
+func replaceMurRefs(t *testing.T, awaitility *Awaitility, murName string, refs []string) {
+	masterUserRecord, err := awaitility.Host().WaitForMasterUserRecord(murName)
+	require.NoError(t, err)
+	var namespaces []toolchainv1alpha1.NSTemplateSetNamespace
+	masterUserRecord.Spec.UserAccounts[0].Spec.NSTemplateSet.ClusterResources = nil
+	for _, ref := range refs {
+		if strings.Contains(ref, "clusterresources") {
+			masterUserRecord.Spec.UserAccounts[0].Spec.NSTemplateSet.ClusterResources = &toolchainv1alpha1.NSTemplateSetClusterResources{
+				TemplateRef: ref,
+			}
+		} else {
+			namespaces = append(namespaces, toolchainv1alpha1.NSTemplateSetNamespace{
+				TemplateRef: ref,
+			})
+		}
+	}
+	masterUserRecord.Spec.UserAccounts[0].Spec.NSTemplateSet.Namespaces = namespaces
+
+	err = awaitility.Client.Update(context.TODO(), masterUserRecord)
+	require.NoError(t, err)
 }
 
 func TestUpdateNSTemplateTier(t *testing.T) {
@@ -160,57 +345,4 @@ func TestUpdateNSTemplateTier(t *testing.T) {
 	// and there should be `MaxPoolSize` TemplateUpdateRequests created
 	err = hostAwaitility.WaitForTemplateUpdateRequests(hostAwaitility.Ns, MaxPoolSize)
 	require.NoError(t, err)
-}
-
-func TestTierTemplates(t *testing.T) {
-	// given
-	tierList := &toolchainv1alpha1.NSTemplateTierList{}
-	ctx, awaitility := testsupport.WaitForDeployments(t, tierList)
-	defer ctx.Cleanup()
-	// when the tiers are created during the startup then we can verify them
-	allTiers := &toolchainv1alpha1.TierTemplateList{}
-	err := awaitility.Client.List(context.TODO(), allTiers, client.InNamespace(awaitility.HostNs))
-	// verify that we have 11 tier templates (4+4+3)
-	require.NoError(t, err)
-	assert.Len(t, allTiers.Items, 11)
-}
-
-func TestUpdateOfNamespacesWithLegacyLabels(t *testing.T) {
-	// given
-	tierList := &toolchainv1alpha1.NSTemplateTierList{}
-	ctx, awaitility := testsupport.WaitForDeployments(t, tierList)
-	defer ctx.Cleanup()
-	for _, nsType := range []string{"code", "dev", "stage"} {
-		err := awaitility.Client.Create(context.TODO(), &corev1.Namespace{
-			ObjectMeta: metav1.ObjectMeta{
-				Name: "legacy-" + nsType,
-				Labels: map[string]string{
-					"toolchain.dev.openshift.com/provider": "codeready-toolchain",
-					"toolchain.dev.openshift.com/owner":    "legacy",
-					"toolchain.dev.openshift.com/tier":     "basic",
-					"toolchain.dev.openshift.com/type":     nsType,
-				},
-			},
-		}, &test.CleanupOptions{})
-		require.NoError(t, err)
-	}
-
-	// when
-	legacySignup := createAndApproveSignup(t, awaitility, "legacy")
-
-	// then
-	verifyResourcesProvisionedForSignup(t, awaitility, legacySignup, "basic")
-}
-
-func newChangeTierRequest(namespace, tier, murName string) *toolchainv1alpha1.ChangeTierRequest {
-	return &toolchainv1alpha1.ChangeTierRequest{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace:    namespace,
-			GenerateName: "changetierrequest-",
-		},
-		Spec: toolchainv1alpha1.ChangeTierRequestSpec{
-			TierName: tier,
-			MurName:  murName,
-		},
-	}
 }
